+++ packages/api/src/praise/controllers.ts
@@ -16,16 +16,9 @@ import {
   TypedRequestQuery,
   TypedResponse,
 } from '@shared/types';
-import { UserModel } from '@user/entities';
-import { UserDocument } from '@user/types';
-import { UserAccountModel } from '@useraccount/entities';
-import { Request, Response } from 'express';
-import { Parser } from 'json2csv';
+import { Request } from 'express';
 import { PraiseModel } from './entities';
-import {
-  calculateDuplicateScore,
-  praiseDocumentTransformer,
-} from './transformers';
+import { praiseDocumentTransformer } from './transformers';
 import {
   PraiseAllInput,
   PraiseDetailsDto,
@@ -137,177 +130,4 @@ export const quantify = async (
   await praise.save();
   const response = await praiseDocumentTransformer(praise);
   res.status(200).json(response);
-};
-
-interface PraiseDtoExtended extends PraiseDto {
-  receiverUserDocument?: UserDocument;
-  giverUserDocument?: UserDocument;
-}
-
-/**
- * //TODO add descriptiom
- */
-export const exportPraise = async (
-  req: TypedRequestBody<QueryInput>,
-  res: Response
-): Promise<void> => {
-  if (!req.query.periodStart || !req.query.periodEnd) {
-    throw new BadRequestError(
-      'You need to specify start and end date for period.'
-    );
-  }
-
-  const praises = await PraiseModel.find({
-    createdAt: {
-      $gt: new Date(String(req.query.periodStart)),
-      $lte: new Date(String(req.query.periodEnd)),
-    },
-  }).populate('giver receiver');
-
-  const quantificationsColumnsCount = parseInt(
-    String(process.env.PRAISE_QUANTIFIERS_PER_PRAISE_RECEIVER)
-  );
-
-  const docs: PraiseDetailsDto[] = [];
-  if (praises) {
-    for (const praise of praises) {
-      const pws: PraiseDtoExtended = await praiseWithScore(praise);
-
-      const receiver = await UserModel.findById(pws.receiver.user);
-      if (receiver) {
-        pws.receiverUserDocument = receiver;
-      }
-
-      if (pws.giver && pws.giver.user) {
-        const giver = await UserModel.findById(pws.giver.user);
-        if (giver) {
-          pws.giverUserDocument = giver;
-        }
-      }
-
-      pws.quantifications = await Promise.all(
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        pws.quantifications.map(async (q: any) => {
-          const quantifier = await UserModel.findById(q.quantifier._id);
-          const account = await UserAccountModel.findOne({
-            user: q.quantifier._id,
-          });
-
-          q.quantifier = quantifier;
-          q.account = account;
-
-          if (q.duplicatePraise) {
-            const praise = await PraiseModel.findById(q.duplicatePraise._id);
-            if (praise && praise.quantifications) {
-              const quantification = praise.quantifications.find((q) =>
-                q.quantifier.equals(q.quantifier)
-              );
-              if (quantification) {
-                q.score = quantification.dismissed
-                  ? 0
-                  : await calculateDuplicateScore(quantification);
-              }
-            }
-          }
-
-          return q;
-        })
-      );
-
-      docs.push(pws);
-    }
-  }
-
-  const fields = [
-    {
-      label: 'ID',
-      value: '_id',
-    },
-    {
-      label: 'DATE',
-      value: 'createdAt',
-    },
-    {
-      label: 'TO USER ACCOUNT',
-      value: 'receiver.name',
-    },
-    {
-      label: 'TO ETH ADDRESS',
-      value: 'receiverUserDocument.ethereumAddress',
-    },
-    {
-      label: 'FROM USER ACCOUNT',
-      value: 'giver.name',
-    },
-    {
-      label: 'FROM ETH ADDRESS',
-      value: 'giverUserDocument.ethereumAddress',
-    },
-    {
-      label: 'REASON',
-      value: 'reason',
-    },
-    {
-      label: 'SOURCE ID',
-      value: 'sourceId',
-    },
-    {
-      label: 'SOURCE NAME',
-      value: 'sourceName',
-    },
-  ];
-
-  for (let index = 0; index < quantificationsColumnsCount; index++) {
-    const quantObj = {
-      label: `SCORE ${index + 1}`,
-      value: `quantifications[${index}].score`,
-    };
-
-    fields.push(quantObj);
-  }
-
-  for (let index = 0; index < quantificationsColumnsCount; index++) {
-    const quantObj = {
-      label: `DUPLICATE ID ${index + 1}`,
-      value: `quantifications[${index}].duplicatePraise`,
-    };
-
-    fields.push(quantObj);
-  }
-
-  for (let index = 0; index < quantificationsColumnsCount; index++) {
-    const quantObj = {
-      label: `DISMISSED ${index + 1}`,
-      value: `quantifications[${index}].dismissed`,
-    };
-
-    fields.push(quantObj);
-  }
-
-  for (let index = 0; index < quantificationsColumnsCount; index++) {
-    const quantUserUsernameObj = {
-      label: `QUANTIFIER ${index + 1} USERNAME`,
-      value: `quantifications[${index}].account.name`,
-    };
-
-    fields.push(quantUserUsernameObj);
-
-    const quantUserEthAddressObj = {
-      label: `QUANTIFIER ${index + 1} ETH ADDRESS`,
-      value: `quantifications[${index}].quantifier.ethereumAddress`,
-    };
-
-    fields.push(quantUserEthAddressObj);
-  }
-
-  fields.push({
-    label: 'AVG SCORE',
-    value: 'score',
-  });
-
-  const json2csv = new Parser({ fields: fields });
-  const csv = json2csv.parse(docs);
-
-  res.attachment('data.csv');
-  res.status(200).send(csv);
 };
